- Version 1.0.4

+++ Sources/Cocoa/TCChatPage.m
@@ -48,7 +48,7 @@
 
 - (void)setFrame:(NSRect)rect
 {
-	float delta = rect.size.width - self.frame.size.width;
+	CGFloat delta = rect.size.width - self.frame.size.width;
 	
 	// Update items
 	float	current_y = 0;
